// frontend/src/generated/models/userAccountInfo.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UserAccountInfo {
  uuid?: string;
  username?: string;
  email?: string;
  roles?: string[];
  permissions?: string[];
  customPermissions?: string[];
  active?: boolean;
  tenantId?: string;
  tenantName?: string;
  allPermissions?: string[];
}